fix size_t on functions

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: gaducurt <gaducurt@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 15:36:02 by gaducurt          #+#    #+#             */
-/*   Updated: 2024/11/09 15:41:56 by gaducurt         ###   ########.fr       */
+/*   Updated: 2024/11/09 16:43:04 by gaducurt         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlen(const char *s)
+#include <libft.h>
+
+size_t	ft_strlen(const char *s)
 {
 	int	i;
